requirements.txt
asgiref==3.2.5
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.4
django-allauth==0.41.0
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-modeltranslation==0.14.4
django-recaptcha3==0.2.0
idna==2.9
oauthlib==3.1.0
Pillow==7.0.0
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
